+++ bot/utils/pavlov.py
@@ -1,9 +1,11 @@
 import logging
+from typing import Optional, Union
 
 import discord
+from discord.ext import commands
+from pavlov import PavlovRCON
 
 from bot.utils import servers, user_action_log
-from pavlov import PavlovRCON
 
 RCON_TIMEOUT = 60
 
@@ -39,8 +41,9 @@ async def check_perm_admin(
             f"ADMIN CHECK FAILED server={server_name}, global_check={global_check}",
             log_level=logging.WARNING,
         )
-        if not ctx.batch_exec:
-            await ctx.send(embed=discord.Embed(description=f"This command is only for Admins."))
+        if hasattr(ctx, "batch_exec"):
+            if not ctx.batch_exec:
+                await ctx.send(embed=discord.Embed(description=f"This command is only for Admins."))
     return False
 
 
@@ -82,12 +85,13 @@ async def check_perm_moderator(
                 f"MOD CHECK FAILED server={server_name}, global_check={global_check}",
                 log_level=logging.WARNING,
             )
-            if not ctx.batch_exec:
-                await ctx.send(
-                    embed=discord.Embed(
-                        description=f"This command is only for Moderators and above."
+            if hasattr(ctx, "batch_exec"):
+                if not ctx.batch_exec:
+                    await ctx.send(
+                        embed=discord.Embed(
+                            description=f"This command is only for Moderators and above."
+                        )
                     )
-                )
         return False
     return True
 
@@ -101,20 +105,23 @@ async def check_perm_captain(ctx, server_name: str = None, global_check: bool = 
             f"CAPTAIN CHECK FAILED server={server_name} global_check={global_check}",
             log_level=logging.WARNING,
         )
-        if not ctx.batch_exec:
-            await ctx.send(
-                embed=discord.Embed(description=f"This command is only for Captains and above.")
-            )
+        if hasattr(ctx, "batch_exec"):
+            if not ctx.batch_exec:
+                await ctx.send(
+                    embed=discord.Embed(description=f"This command is only for Captains and above.")
+                )
         return False
     return True
 
 
-async def exec_server_command(ctx, server_name: str, command: str, polling=False):
+async def exec_server_command(
+    ctx: Optional[Union[commands.Context, PavlovRCON]], server_name: str, command: str
+) -> [dict, Optional[PavlovRCON]]:
     pavlov = None
-    if not polling:
+    if ctx is not None and isinstance(ctx, commands.Context):
         if hasattr(ctx, "pavlov"):
             pavlov = ctx.pavlov.get(server_name)
-        if not hasattr(ctx, "pavlov") or pavlov is None:
+        if pavlov is None:
             server = servers.get(server_name)
             pavlov = PavlovRCON(
                 server.get("ip"),
@@ -127,18 +134,16 @@ async def exec_server_command(ctx, server_name: str, command: str, polling=False
             else:
                 ctx.pavlov[server_name] = pavlov
         data = await pavlov.send(command)
-        return data
-    else:
-        if ctx is not None:
-            pavlov = ctx
-        elif ctx is None:
-            server = servers.get(server_name)
-            pavlov = PavlovRCON(
-                server.get("ip"),
-                server.get("port"),
-                server.get("password"),
-                timeout=RCON_TIMEOUT,
-            )
-        ctx = pavlov
-        data = await pavlov.send(command)
         return data, ctx
+    if ctx is None:
+        server = servers.get(server_name)
+        pavlov = PavlovRCON(
+            server.get("ip"),
+            server.get("port"),
+            server.get("password"),
+            timeout=RCON_TIMEOUT,
+        )
+    elif isinstance(ctx, PavlovRCON):
+        pavlov = ctx
+    data = await pavlov.send(command)
+    return data, pavlov